fix inventory handling

+++ src/net/packet/InventoryAckPacket.cs
@@ -2,18 +2,18 @@ namespace BlockGame.net.packet;
 
 /** S→C: 0x43 - confirm or reject inventory transaction */
 public struct InventoryAckPacket : Packet {
-    public byte invID;
+    public int invID;
     public ushort actionID;
     public bool acc;
 
     public void write(PacketBuffer buf) {
-        buf.writeByte(invID);
+        buf.writeInt(invID);
         buf.writeUShort(actionID);
         buf.writeBool(acc);
     }
 
     public void read(PacketBuffer buf) {
-        invID = buf.readByte();
+        invID = buf.readInt();
         actionID = buf.readUShort();
         acc = buf.readBool();
     }
